Cooked The Sim More (fast) + View Sim on Edit Case Page
Allows you to view the simulation for a case on the Edit Case Page - Even if that sim hasn't been generated before. Some cases don't have the correct timestamp form so it will work for new cases and if you add a new case, the format should be correct so then it work for that new case.